+++ engine/Components/CreateElements.py
@@ -11,7 +11,7 @@ def create_array_of_elements(statement: str) -> list:
 	:param statement: string
 	:return: Returns an array of elements
 	"""
-	logger.info("Creating an array from a string...")
+	logger.debug("Creating an array from a string...")
 	logger.debug("Calling create_array_of_elements(%s)", statement)
 
 	array_of_elements = []
@@ -20,5 +20,5 @@ def create_array_of_elements(statement: str) -> list:
 		array_of_elements.append(char)
 
 	logger.debug("array_of_elements=%s", array_of_elements)
-	logger.info("Finished creating array.")
+	logger.debug("Finished creating array.")
 	return array_of_elements