+++ engine/Computation/TableMatrix/RecursiveTable.py
@@ -13,8 +13,7 @@ def recursive_table(n):
 	:param n: (int) Number of variables
 	:return: Returns an array table with T/F values
 	"""
-	# logger.info("Calling a recursive function to create entry values for matrix...")
-	logger.info("Creating table entries recursively...")
+	logger.debug("Creating table entries recursively...")
 
 	logger.debug("n=%d", n)
 	array_table = []
@@ -22,7 +21,6 @@ def recursive_table(n):
 		array_table = [TRUE, FALSE]
 	elif n != 1:
 		array_table = recursive_table(n - 1)
-		# temp_table = []
 		temp_table = array_table.copy()
 		for j in range(0, int(2 ** n / 2)):
 			if j % 2 == 0:  # even
@@ -38,5 +36,5 @@ def recursive_table(n):
 				array_table[i] = str(array_table[i]) + FALSE
 
 	logger.debug("array_table=%s", array_table)
-	logger.info("Finished creating array of table entries.")
+	logger.debug("Finished creating array of table entries.")
 	return array_table